Time: 3 ms (77.87%), Space: 69.3 MB (87.82%) - LeetHub

+++ 0643-maximum-average-subarray-i/0643-maximum-average-subarray-i.java
@@ -1,16 +1,16 @@
 class Solution {
     public double findMaxAverage(int[] nums, int k) {
         double avg=0;
-        int sum=0;
+        double sum=0;
         for(int i=0;i<k;i++){
             sum+=nums[i];
         }
-        double max=(double)sum/k;
+        double max=sum/k;
 
         for(int i=k;i<nums.length;i++){
             sum+=nums[i];
             sum-=nums[i-k];
-            avg=(double)sum/k;
+            avg=sum/k;
             if(max<avg){
                 max=avg;
             }
